requirements.txt
asgiref==3.3.1
Django==3.1.4
django-phonenumber-field==5.0.0
django-tastypie==0.14.3
phonenumbers==8.12.15
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
